# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OnlinePort",
        "version": "0.1"
    },
    "paths": {
...
        "/api/containers": {
            "get": {
                "tags": [
                    "Containers"
                ],
                "summary": "Get list of containers",
                "operationId": "16104655b4f5d229e273970e21bafd3a",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "post": {
                "tags": [
                    "Containers"
                ],
                "summary": "Store Container",
                "operationId": "4c45cd86d437e437ec13145bcc70277a",
                "parameters": [
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ContainerNameTest"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "10 тонн"
                                    },
                                    "kit": {
                                        "type": "string",
                                        "example": "kit"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
        "/api/containers/{id}": {
            "get": {
                "tags": [
                    "Containers"
                ],
                "summary": "Show Container",
                "description": "Show Container",
                "operationId": "083e41584b211785d18e9a8261c92e0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "put": {
                "tags": [
                    "Containers"
                ],
                "summary": "Update Containers",
                "description": "Update Containers",
                "operationId": "Update Containers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    },
                    {
                        "name": "X-Localization",
                        "in": "header",
                        "description": "Localization",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ru": {
                                "summary": "Russian",
                                "value": "ru"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ContainerNameTestUpdate"
                                    },
                                    "weight": {
                                        "type": "string",
                                        "example": "10 тонн"
                                    },
                                    "kit": {
                                        "type": "string",
                                        "example": "kit"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Containers"
                ],
                "summary": "Summary",
                "description": "Containers",
                "operationId": "Delete container",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SUCCESS"
                    }
                }
            }
        },
...
    }
}